baseInfo url

+++ routers/baseInfo.js
@@ -5,6 +5,7 @@ const multer = require('koa-multer');
 const upload = multer({ dest: path.join(__dirname, '../assets/images') });
 const fileOperation = require('./common.js').fileOperation;
 const loggerError = require('./common.js').loggerError;
+const defaultUrl = path.join(__dirname, `../assets/images/baseInfo/pic01.jpg`);
 module.exports = (db) => {
     router.get('/', async(ctx) => {
         try {
@@ -44,11 +45,13 @@ module.exports = (db) => {
                 });
                 if (logoUrl) {
                     let tmp_path = path.join(__dirname, `../assets${logoUrl}`);
-                    await fs.unlink(tmp_path, (err) => {
-                        if (err) {
-                            throw `error with unlink imageFile:${err}`;
-                        }
-                    });
+                    if(tmp_path != defaultUrl){
+                        await fs.unlink(tmp_path, (err) => {
+                            if (err) {
+                                throw `error with unlink imageFile:${err}`;
+                            }
+                        });
+                    }
                 }
                 return ctx.body = {
                     status: 200,
